refactor(observability): sweep gateway modules onto domain telemetry imports
