+++ surfsense_backend/app/gateway/whatsapp/translator_baileys.py
@@ -10,7 +10,7 @@ from app.gateway.base.adapter import BasePlatformAdapter, PlatformSendResult
 from app.gateway.base.formatting import split_text_message
 from app.gateway.base.translator import BaseStreamTranslator, GatewayStreamEvent
 from app.gateway.whatsapp.adapter_baileys import WhatsAppBaileysAdapter
-from app.observability.metrics import (
+from app.observability.domains.gateway import (
     record_gateway_hitl_aborted,
     record_gateway_outbound,
 )
